Unify pagination design
