+++ apps/web/src/app/(default)/admin/brukere/user-table.tsx
@@ -1,41 +1,139 @@
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuLabel,
-  DropdownMenuSeparator,
-  DropdownMenuTrigger,
-} from "@radix-ui/react-dropdown-menu";
-import { RxDotsHorizontal as Dots } from "react-icons/rx";
+"use client";
+
+import { useEffect, useMemo, useState } from "react";
+import { RxArrowLeft, RxArrowRight } from "react-icons/rx";
 
 import { type Group } from "@echo-webkom/db/schemas";
 
-import { Button } from "@/components/ui/button";
-import { TableCell, TableRow } from "@/components/ui/table";
+import { Container } from "@/components/container";
+import { Heading } from "@/components/typography/heading";
+import { HoldableButton } from "@/components/ui/holdable-button";
+import { Input } from "@/components/ui/input";
+import { Label } from "@/components/ui/label";
+import { Table, TableBody, TableHead, TableHeader, TableRow } from "@/components/ui/table";
 import { type AllUsers } from "./page";
-import { UserForm } from "./user-form";
+import { UserTable } from "./user-table-row";
 
-type User = AllUsers[number];
+/**
+ * Component used to display a table of users.
+ * Created to be client sided.
+ */
+export const UserTableView = ({ users, groups }: { users: AllUsers; groups: Array<Group> }) => {
+  const [searchQuery, setSearchQuery] = useState("");
+  const [page, setPage] = useState(1);
+  const [pageSize, setPageSize] = useState(10);
 
-export const UserTable = ({ user, groups }: { user: User; groups: Array<Group> }) => {
+  // Filter users based on search query
+  const filteredUsers = (users ?? [])?.filter(
+    (user) =>
+      (user.name?.toLowerCase().includes(searchQuery.toLowerCase()) ?? false) ||
+      user.email.toLowerCase().includes(searchQuery.toLowerCase()),
+  );
+
+  useEffect(() => {
+    setPage(1);
+  }, [searchQuery, pageSize]);
+  const totalPages = Math.ceil(filteredUsers.length / pageSize) || 1;
+
+  const handleNextPage = () => {
+    if (page === totalPages) return;
+    setPage(page + 1);
+  };
+
+  const handlePreviousPage = () => {
+    if (page === 1) return;
+    setPage(page - 1);
+  };
+
+  const handleHoldPreviousPage = () => {
+    if (page === 1) return;
+    setPage(1);
+  };
+
+  const handleHoldNextPage = () => {
+    if (page === totalPages) return;
+    setPage(totalPages);
+  };
+
+  if (page > totalPages) {
+    setPage(totalPages);
+  }
+
+  const paginatedUsers = useMemo(
+    () => filteredUsers.slice((page - 1) * pageSize, page * pageSize),
+    [filteredUsers, page, pageSize],
+  );
+
   return (
-    <TableRow key={user.id}>
-      <TableCell>{user.name}</TableCell>
-      <TableCell>{user.email}</TableCell>
-      <TableCell>
-        <DropdownMenu>
-          <DropdownMenuTrigger asChild>
-            <Button variant="ghost" className="h-8 w-8 p-0">
-              <span className="sr-only">Open menu</span>
-              <Dots className="h-4 w-4" />
-            </Button>
-          </DropdownMenuTrigger>
-          <DropdownMenuContent align="end">
-            <DropdownMenuLabel>Gjør endringer</DropdownMenuLabel>
-            <DropdownMenuSeparator />
-            <UserForm user={user} groups={groups} />
-          </DropdownMenuContent>
-        </DropdownMenu>
-      </TableCell>
-    </TableRow>
+    <Container>
+      <Heading className="mb-4">Brukere</Heading>
+
+      <div className="flex flex-col gap-2">
+        <div className="flex justify-between gap-2">
+          <div className="mt-auto flex items-center gap-2">
+            <HoldableButton
+              variant="outline"
+              size="icon"
+              onClick={handlePreviousPage}
+              disabled={page === 1}
+              onHold={handleHoldPreviousPage}
+            >
+              <span className="sr-only">Tilbake</span>
+              <RxArrowLeft className="h-4 w-4" />
+            </HoldableButton>
+            <span>
+              {page} av {totalPages}
+            </span>
+            <HoldableButton
+              variant="outline"
+              size="icon"
+              onClick={handleNextPage}
+              onHold={handleHoldNextPage}
+              disabled={page === totalPages}
+            >
+              <span className="sr-only">Neste</span>
+              <RxArrowRight className="h-4 w-4" />
+            </HoldableButton>
+          </div>
+
+          <div className="flex items-center gap-2">
+            <div className="flex max-w-[75px] flex-col gap-2">
+              <Label>Antall:</Label>
+              <Input
+                type="number"
+                value={pageSize}
+                onChange={(e) => setPageSize(parseInt(e.target.value))}
+              />
+            </div>
+
+            <div className="flex flex-col gap-2">
+              <Label>Søk:</Label>
+              <Input
+                value={searchQuery}
+                onChange={(e) => setSearchQuery(e.target.value)}
+                placeholder="Søk..."
+              />
+            </div>
+          </div>
+        </div>
+
+        <Table>
+          <TableHeader>
+            <TableRow>
+              <TableHead>Navn</TableHead>
+              <TableHead>E-post</TableHead>
+              <TableHead>Endre</TableHead>
+            </TableRow>
+          </TableHeader>
+          <TableBody>
+            {paginatedUsers.map((user) => (
+              <UserTable key={user.id} user={user} groups={groups} />
+            ))}
+          </TableBody>
+        </Table>
+
+        <span className="font-medium">Totalt: {filteredUsers.length} brukere</span>
+      </div>
+    </Container>
   );
 };
